model-server/requirements.txt
flask
openai
python-dotenv
numpy
pandas
torch
transformers
accelerate
bitsandbytes
sentencepiece
neo4j
requests
beautifulsoup4
